dev: add development container to docker-compose.yaml
To simplify contribution.
To install the dependencies:
docker compose run --rm php composer install --prefer-dist

To run the tests:
docker compose run --rm php vendor/bin/phpunit --verbose

The profile in the docker-compose.yaml file is used that
the php service does not start with the mysql service,
and it is only used for one shot commands, not as
a daemon.